refactor: main.js
add app.js, install packages intended for tests
