+++ miamoremustachio/ToDo/.gitignore
@@ -1,3 +1,4 @@
 node_modules
 .env
 migrate-mongo-config.js
+jsconfig.json
